requirements.txt
beautifulsoup4==4.8.1
requests==2.22.0
ImageHash==4.0
Pillow==6.2.1
psycopg2-binary==2.8.4
PyYAML==5.2
